implement dry_run option for OacisCli#show_host

+++ lib/cli/oacis_cli_common.rb
@@ -38,6 +38,7 @@ EOS
     hosts = Host.all.map do |host|
       {id: host.id.to_s, name: host.name, hostname: host.hostname, user: host.user}
     end
+    return if options[:dry_run]
     File.open(options[:output], 'w') {|io|
       io.puts JSON.pretty_generate(hosts)
       io.flush
